refactor (.github/workflows): Dropped support for windows in the workflows
Although I've been trying to add a CI/CD pipeline for both windows and linux, wxWidgets has made it impossible to work with windows in github Actions. Packaging will work the same, but windows will not be built on the same workflow. It will take the package from another repo