refactor\!: dropped compatibility for linux (explanation next)
DeskUp uses X11 as it's backend for linux. Now a days, virtually every linux distro has switched on from X11 to Wayland due to security reasons, among others. Even though it is still possible to use a X11 server inside a Wayland-type distro, less than 10% of the apps that run on them support X11. This means that realistically, DeskUp is unable to run smoothly on linux