chore (desk_up_error): dropped exceptions and created an error type used for silent failing
using exceptions meant blocking the execution of the calls to the backend. The problem was that some operations can be retried and/or skipped while still finishing the call. If a function needs to terminate because it has encountered a fatal error, it can just return after filling the error code. The frontend can collect this information and do the aproppriate thing while knowing that an operation might have gone wrong